+++ app/src/main/java/com/example/electronicdiary/data_classes/Semester.java
@@ -3,11 +3,11 @@ package com.example.electronicdiary.data_classes;
 import org.jetbrains.annotations.NotNull;
 
 public class Semester {
-    private final int id;
+    private final long id;
     private final int year;
     private final boolean isFirstHalf;
 
-    public Semester(int id, int year, boolean isFirstHalf) {
+    public Semester(long id, int year, boolean isFirstHalf) {
         this.id = id;
         this.year = year;
         this.isFirstHalf = isFirstHalf;
@@ -19,7 +19,7 @@ public class Semester {
         return getFirstHalf() + "/2 " + getYear();
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
