+++ app/src/main/java/com/example/electronicdiary/data_classes/StudentPerformanceInModule.java
@@ -2,17 +2,17 @@ package com.example.electronicdiary.data_classes;
 
 public class StudentPerformanceInModule {
     private final int moduleNumber;
-    private final int studentId;
-    private final int groupId;
-    private final int subjectId;
-    private final int lecturerId;
-    private final int seminarianId;
-    private final int semesterId;
+    private final long studentId;
+    private final long groupId;
+    private final long subjectId;
+    private final long lecturerId;
+    private final long seminarianId;
+    private final long semesterId;
 
     private int earnedPoints = -1;
     private boolean isHaveCredit = false;
 
-    public StudentPerformanceInModule(int moduleNumber, int studentId, int groupId, int subjectId, int lecturerId, int seminarianId, int semesterId) {
+    public StudentPerformanceInModule(int moduleNumber, long studentId, long groupId, long subjectId, long lecturerId, long seminarianId, long semesterId) {
         this.moduleNumber = moduleNumber;
         this.studentId = studentId;
         this.groupId = groupId;
@@ -22,7 +22,7 @@ public class StudentPerformanceInModule {
         this.semesterId = semesterId;
     }
 
-    public StudentPerformanceInModule(int moduleNumber, int studentId, int groupId, int subjectId, int lecturerId, int seminarianId, int semesterId, int earnedPoints, boolean isHaveCredit) {
+    public StudentPerformanceInModule(int moduleNumber, long studentId, long groupId, long subjectId, long lecturerId, long seminarianId, long semesterId, int earnedPoints, boolean isHaveCredit) {
         this.moduleNumber = moduleNumber;
         this.studentId = studentId;
         this.groupId = groupId;
@@ -38,27 +38,27 @@ public class StudentPerformanceInModule {
         return moduleNumber;
     }
 
-    public int getStudentId() {
+    public long getStudentId() {
         return studentId;
     }
 
-    public int getGroupId() {
+    public long getGroupId() {
         return groupId;
     }
 
-    public int getSubjectId() {
+    public long getSubjectId() {
         return subjectId;
     }
 
-    public int getLecturerId() {
+    public long getLecturerId() {
         return lecturerId;
     }
 
-    public int getSeminarianId() {
+    public long getSeminarianId() {
         return seminarianId;
     }
 
-    public int getSemesterId() {
+    public long getSemesterId() {
         return semesterId;
     }
