+++ app/src/main/java/com/example/electronicdiary/data_classes/Subject.java
@@ -3,15 +3,15 @@ package com.example.electronicdiary.data_classes;
 import org.jetbrains.annotations.NotNull;
 
 public class Subject {
-    private final int id;
+    private final long id;
     private final String title;
 
-    public Subject(int id, String title) {
+    public Subject(long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
